+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-04-01 21:37:46
-// Commit: test(network): add unit tests for network business logic
+// Updated at: 2026-04-02 09:27:35
+// Commit: feat(network): integrate custom telemetry logger in network
 
 class networkComponentState {
-  final String commitRef = '31';
-  final String lastMessage = 'test(network): add unit tests for network business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-01T21:37:46+05:00');
+  final String commitRef = '41';
+  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
+  final DateTime timestamp = DateTime.parse('2026-04-02T09:27:35+05:00');
 }